cosmo/wakeword/pyproject.toml
[project]
name = "cosmo-wakeword"
version = "0.1.0"
description = "Wakeword detection service for Cosmo"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openwakeword",
    "numpy",
    "sounddevice",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
